move apiLoggingConfiguration to logger.configuration & fix code lint issues
