+++ src/main/java/io/apimatic/core/logger/ConsoleLogger.java
