+++ src/main/java/io/apimatic/core/logger/LoggerFactory.java
@@ -4,12 +4,18 @@ import io.apimatic.coreinterfaces.logger.ApiLogger;
 import io.apimatic.coreinterfaces.logger.configuration.ReadonlyLoggingConfiguration;
 
 public class LoggerFactory {
+
+    /**
+     * Private constructor to prevent instantiation
+     */
+    private LoggerFactory() {
+    }
+
     /**
      * Returns an instance of {@link ApiLogger} based on the provided
      * {@code ReadonlyLogging} configuration. If the logger in the configuration is
      * null, it returns an instance of {@code NullSdkLogger}. Otherwise, it returns
      * an instance of {@code SdkLogger} initialized with the provided configuration.
-     *
      * @param config the {@link ReadonlyLoggingConfiguration} configuration to be
      *               used for creating the logger
      * @return an instance of {@link ApiLogger} based on the provided configuration
